pyproject.toml
[project]
name = "streamlit-qa-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.0.13",
    "docx2txt>=0.9",
    "langchain>=0.3.26",
    "langchain-community>=0.3.26",
    "langchain-groq>=0.3.4",
    "langchain-ollama>=0.3.3",
    "pypdf>=5.7.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.46.1",
    "tiktoken>=0.9.0",
]
